trying to write test cases for UserGeneratorControllerTest
