+++ src/main/java/useCases/ReadingLogger.java
@@ -26,4 +26,3 @@ public class ReadingLogger {
         readingLog.addToLog(reading);
     }
 }
-
